+++ src/main/java/org/greencodeinitiative/creedengo/python/PythonRuleRepository.java
@@ -41,6 +41,7 @@ public class PythonRuleRepository implements RulesDefinition, PythonCustomRuleRe
             AvoidListComprehensionInIterations.class,
             DetectUnoptimizedImageFormat.class,
             AvoidMultipleIfElseStatementCheck.class,
+            PandasRequireUsecolsArgument.class,
             OptimizeSquareComputation.class
     );